apps/api/pyproject.toml
[project]
name = "greenwich-smartfaq-api"
version = "0.1.0"
description = "FastAPI backend for Greenwich SmartFAQ"
authors = [{ name = "Dev Team" }]
requires-python = ">=3.11"
dependencies = [
  # API
  "fastapi[standard]~=0.120",
  "uvicorn[standard]~=0.30",

  # DB & migrations
  "sqlalchemy~=2.0",
  "psycopg[binary]~=3.2",
  "alembic~=1.13",

  # Settings
  "pydantic-settings~=2.4",

  # Auth
  "python-jose[cryptography]~=3.3",
  "passlib[bcrypt]~=1.7",

  # LangChain 1.0 + integrations
  "langchain>=1,<2",
  "langchain-openai>=0.2,<1",
  "langchain-chroma>=0.1.2",

  # Embeddings / LLM
  "sentence-transformers~=3.0",
  "openai>=1.40,<2.0",

  # Form, Task, Cache
  "python-multipart~=0.0.9",
  "celery~=5.4",
  "redis~=5.0",
]

[project.optional-dependencies]
dev = [
  "pytest~=8.3",
  "pytest-asyncio~=0.24",
  "pytest-cov~=5.0",
  "httpx~=0.27",
  "ruff~=0.5",
  "black~=24.8",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.black]
target-version = ["py311"]
line-length = 100
